Hack to make tasks work for now
Due to 2 celery App instances being created (one for dashboard, one for
websecmap) during loading of worker some tasks in dashboard get lost as
the autodiscover only works on a specific app. Importing the module with
tasks explicitly loads the tasks for both App instances, fixing the
problem for now while we investigate how to resolve this properly.